Restore some of the startup parameters closer to what Quax originally had.
This has a bit more trouble starting a hard drive motor, but is quite a
bit smoother on stronger motor. The original tweaking was done to work
around the lack of a deadband between valid signals and off and threshold
between start and stop that was letting PPM jitter interfere with the
startup procedure.