Fix the throttle handling to have a band between valid and motor-stop.
Previously, the code relied on MIN_DUTY to offset a motor start from the
shortest valid pulse length, but we have MIN_DUTY set to 1, so there was
only a band of a few microseconds where it actually worked.

While we're at it, add an anti-jitter threshold to the motor startup,
which seemed to be causing quite a bit of the startup issues in my test cases.