Set/clear FULL_POWER before POWER_OFF.
Previously, an interrupt that comes along within set_new_duty might see
FULL_POWER and POWER_OFF set at the same time, which could cause it to
both skip the ON cycle nFET advance due to POWER_OFF and also skip
turning off of the nFET due to FULL_POWER, which will stick it on for an
extra whole duty cycle. This should be nicer.