Set OSCCAL to 0xff to run all boards at ~16MHz.
This should be safe if we do not write to the EEPROM or Flash while
OSCCAL is 0xff. If we ever want to use these, we should be able to
restore it to 0x9f. Meanwhile, this should let us be compatible with
boards with 16MHz oscillators with no build/code changes, and lets
us have an even wider duty cycle range with twice the steps. Woot.