Always update duty cycle with timing updates.
Without this, if we lose the RC pulse, we will actually never decrease
the duty cycle until the RC pulse timeout happens, which might take 1-2
seconds! This is bad, mmmkay. If the motor is planted inthe ground, we
really want to limit the duty cycle immediately (preferrably to 0, but
for now to PWR_MAX_RPM1).