Fix the 24-bit timing code to have no misses at the low end and near 0 aliases in 16-bit space.
With the introduction of the 24-bit extension, perfect timing is critical
because of all of the 24-bit possibilities where the 16-bit values are
around 0. This code was unit-tested from 65530 to 65550 and behaves well,
so there should be no case where the timer misses or fires early. With
clk/8 instead of clk/1, it should now be safe to change the adiw YL, 7 to
adi YL, 1.