Add RC-car style reverse throttle support (but not yet the commutation), cooperative scheduling of duty cycle update during wait periods, and start support for i2c/twi input and programmable throttle calibration.
This is more of a framework change than really support for any of this,
yet, but now the rest should be possible.

Duty cycle updates should be now around 32us average, 65us maximum (just
the PWM frequency and processing overhead), rather than being aligned to
commutation timing as it was previously. This could reduce update lag
significantly, depending on motor speed. At first I tried processing in
the interrupt with re-entant preemption points, but this still introduced
audible PWM noise and started to get pretty ugly. The trade-off is two
more instructions in all of the spin-wait places, and some clobbering
avoidance.

Also added is support for limited throttle jump distance. The idea here
to keep a floating limit significantly above the current requested duty,
and do not allow jumping beyond that except by limited steps applied at
each full commutation cycle. This was the result after tring but failing
to do it by timing alone. Without exact knowledge of the motor and
operating voltage, I don't think it's possible to do a reasonable limit
without some kind of timing change feedback, and doing that requires a
delay anway, so we're back where we started. Off by default, but it may
make sense to enable it on the boards that reset if overcurrented easily.