Use F_CPU and CPU_MHZ to support different clock rates.
Any board type can be changed to a different F_CPU, and we will
set OSCCAL appropriately. If there is an external resonator and
fuses are set to use it, OSCCAL will have no effect and F_CPU
must match the resonator speed.

Note that the POWER_RANGE is not adjusted automatically, so 8MHz
boards will get ~9kHz PWM by default.