Fix regression since 8a39bb46c (24-bit timing) which causes motor timing loss (usually a stop and restart).
If enough interrupts are pending or come up between the read of TCNT1L/H
and the post sampling of tcnt1x (while interrupts are enabled), tcnt1x
may not have been updated yet because of the way interrupts are
implemented on the AVR (one non-interrupt instruction is always executed
between interrupts). This means that if timer1 wrapped before TCNT1L/H
was read, tcnt1x may not be updated in time and the copy to temp3 would
be low by one, introducing a timing skew of -65536 cycles.

To fix this, we read tcnt1x and TIFR in the critical section and then do
a correction afterwards, similar to the RC pulse interrupt code.