switch_power_off was racy in the same way that 36091ee57 fixes.
TOV2 could stay set for a few more instructions after we've stopped the
timer. The interrupt could eventually fire after we've turned off the
FETs and assumed they will stay off.

We could just disable the interrupt vector in TIMSK, but then we would
have to enable it again, too, so it is less instructions to just clear
the pending interrupt after stopping the timer, then turn off all FETs.

This would have been _very_ hard to hit, since the timer2 overflow
interrupt is second in priority only to the ext_int0 in our use, so
it could probably only happen if high frequency (>200kHz) is being
fed to the RC pulse input pin, and then only if unlucky enough to
be in the right commutation stage for it to be matter. This is
impossible since 4a260627.