Double MAX_DRIFT_PULS and wait after detected movement during calibration.
This results from testing calibration on a KK board (with no oscillator)
and Turnigy Plushes (with no oscillator) that resulted in quite a bit of
jitter, causing calibration times to be a bit unpredictable.

Previously, MAX_DRIFT_PULS was fairly low because the tail end of input
changing could be accumulated as part of the average. Instead, we now
call wait240ms whenever movement is detected, then start relearning.
This should be more accurate and still work if the input is jittery.