Adjust PROGRAM_RC_PULS and use a different value when !RC_PULS_REVERSE.
This was a bit high because of not wanting to get in the way of neutral
in the RC_PULS_REVERSE case. Instead, we now use about the same level in
this case, and just under "half stick" (1460us) in the normal case. This
should allow calibration compatibility with the Free Flight Controller,
which has a maximum output pulse of 1640us. I am not sure if it has a
calibration output mode, however.