Remove accidental lossy cryptography from EEPROM routines.
The old routines were clobbering the address register with the data, and
so occasionally "forgot" calibration or locked up on boot when the bytes
data bytes matched or caused a loop. :(

While we're at it, collapse the loop into one function, which saves a
good number of instructions, and makes the register use more obvious.