Enable watchdog support and add some new beeps if it occurs.
The minimum watchdog timeout is kind of too long (16.3ms) to perhaps
prevent damage in the event of a lockup, but it might help. It's
better than without it, I suppose.

A siren sound (low, high beeps) forever indicate that a watchdog reset
has occurred. Also, while testing lockups by touching the resonator,
I was also able to cause resets with all bits clear. So, unknown reset
causes are now beeped with low/high beeps to indicate the bits in MCUCSR,
for debugging.