Add UART control (X3D style).
This is based on and compatible with Bernhard's uart_r06_01.zip at
http://home.versanet.de/~b-konze/blc_6a/blc_6a.htm which seems to
have been implemented for this project:
http://www.control.aau.dk/uav/reports/10gr833/10gr833_student_report.pdf

UART input is enabled on AfroESC boards by default, but also works on
Mystery/BlueSeries-style boards because the PPM input pin is bridged to
the RXD pin. This has been tested on the F-30A, but USE_INT0 should be
disabled to avoid short low pulses being interpreted as idle throttle
pulses, causing early arming and disabling of the UART. We will have to
come up with a better solution for HK programming card support.

It probably makes sense to adjust this protocol to support 16-bit input,
and possibly support Atmel's multi-processor communication mode (MPCM),
which turns the serial input into an addressable bus with hardware
filtering, similar to I2C. This could be an alternative to I2C that works
on some hardware without modification, or on AfroESCs with dedicated UART
pads.