Prevent throttle calibration on brown-out reset.
I was testing a newly-built quad with a NiCd battery that could not give
sufficient current, so all ESCs reset due to brown-out. However, the
flight board did not, causing it to keep sending out a high signal from
the radio, causing one ESC to recalibrate incorrectly. Preventing the
calibration code from working when a brown-out reset is detected prevents
this case.

It is possible to cause a brown-out reset when connecting the battery, if
the connection is not made firmly. Intended calibration should be rare
enough that this is acceptable, and the benefit of avoiding accidental
calibration should be worthwhile.