Add stk500v2 boot loader with Turnigy USB Linker signalling on PWM input.
Fuses must be set to 512-word boot loader (0xe00), and BOOTRST enabled.

This works with the Turnigy USB linker...or mine, at least.

Reading/writing of flash and EEPROM is supported with a avrdude; eg:

	avrdude -p m8 -b 9600 -c stk500v2 -D -P /dev/ttyUSB0 -u <commands>

Chip erase (avrdude -e) does not actually chip erase, but just erases the
EEPROM instead.

Higher signalling rates should work over the wire, but the Turnigy USB
linker only appears to support 9600 baud.

We support flashing over the boot loader by running out of the way if we
are erasing the page that would know how to flash it back after erasing.
There are some nops for future expansion, but significantly changing the
boot loader in-place is tricky. Hopefully this will work for most cases,
and make updates much easier!