Replace .ifdefs with #ifdefs
avra currently does not care, but avrasm2
treats defined() in the assembler macros (.if)
and defined() in preprocess macros (#if) differently.

Assembler definition is:

http://support.atmel.no/knowledgebase/avrstudiohelp/mergedProjects/AVRASM/Html/expressions.html#DEFINED(symbol

DEFINED(symbol) Returns true if symbol is previously defined using
.equ/.set/.def directives. Normally used in conjunction with .if
directives (.if defined(foo)), but may be used in any context. It
differs from other functions in that parentheses around its argument are
not required, and that it only makes sense to use a single symbol as
argument.

But by the time we reach the assembler, our "-D afro2_esc"
symbol gets expanded by the preprocessor to integer 1.

This causes the following error with avrasm2:

tgy.asm(53): Including file 'm8def.inc'
tgy.asm(85): error: syntax error, unexpected INTEGER, expecting SYMBOL

This can be avoided by using -D afro2_esc=SOMETHING but
this in turn is not (yet) supported by avra, which has no
real C preprocessor yet.

AVRASM2 documentation explains further:

http://support.atmel.no/knowledgebase/avrstudiohelp/mergedProjects/AVRASM/Html/AVR_Assembler_2_-_Known_Issues.htm#defined

* The preprocessor 'defined' keyword only relates to symbols defined
with #define, and correctly does this only in preprocessor conditionals
(#if/#elif).

* In all remaining code, the assembler's notion of defined is used,
the correct behaviour would be to only recognize it in assembler
conditionals (.if/.elif).

So the solution is to change assembler conditionals (.if/.endif)
with C preprocessor conditionals (#if/#endif) which will work
with both assemblers correctly.