Inline commutation routines since they all only have one call site.
I went and made a whole com_step loop with inc/dev for forward/reverse,
automatic mux selection, etc., and it ended up being ~30 words smaller,
but added another 1.5 microsecond or so, which is not worth it. We are
already tight on high electrical-RPM motors as it is. Inlining them
instead saves 7 cycles per step and 12 words.

The mux selection will have to be revisited when the ADC is used for
other stuff.