Support assisted jumping to any boot loader, regardless of if the Turnigy USB linker-compatible one was included at build time.
This allows mixing and matching of any boot loader that fits, without
requiring the boot loader to be caught only at power-up.

The included boot loader can be installed once and the main flash area
updated with build without BOOT_LOADER set. As long as BOOT_JUMP is set,
we will jump to a boot loader, if present, when the PWM input pin stays
high.