Move reset code to the end.
This should be safer in the event of an interruped or incomplete flash,
and should allow for more interrupt space that can be reached by Z while
ZH is 0. For bigger chips, we might have to be careful about how far rjmp
can reach, but it's fine on the ATmega8 assuming wrapping is enabled.