Replace timeoutMIN with commutation stretching to combat comparator bias.
In most cases, the comparator output will stay low or high when the motor
is not spinning. Normally this is not a problem as the motor will be
perturbed enough by the next commutation to start to get a reasonable
signal. However, in some cases, the motor may not move enough and we will
just stay in the state of skipping every other commutation until some
external force helps generate enough back-EMF to sway the comparator
output.

Ideally, it would be nice to have an "equal" output from the comparator,
but this is not available.

The start timeout min/max were intended to assist pushing motors that
might otherwise not align without increasing the commutation rate, but
this has never really been the problem. What we really need instead is a
delay for at least long enough to allow the magnetic field to build up
and move the motor before continuing to the next step.

The problem with a fixed START_DELAY_US (as introduced in b86dc7cfadc6)
is that too long a delay will cause problems starting low-mass motors or
motors at high voltage, since we may wait too long and miss the zero
crossing. So, here we introduce a delay which starts at START_DELAY_US
and builds up in steps until we two successful zero crossings. The
default delay wraps around in a byte and will reach a maximum of 2ms.

This should fix the starting of any motor that might sometimes vibrate
or buzz when stopped in certain cogging positions.