Limit zero crossing filter backtracking while running.
The back-EMF from some motors when clipping through the high side body
diode seems to create noise that can significantly delay the ZC filter
and lead to significantly retarded timing that can then blind the next
commutation event due to the blanking period and not completing until
after the zero crossing has already passed. This tends to be a problem
at lower speeds and higher currents.

For example, dragging the bell of the 16 pole Multistar (HengLi) 4822
390KV at about 30% duty cycle at 16V would cause significant current draw
and broken timing. Heavy acceleration would also cause synchronization
loss that even the demag_timeout handler could not help with.

Limiting the backtracking (the amount at which we will "back off") from
an accepted zero crossing point seems to greatly help here. This also
seems to help other cases such as switching directions as quickly as
possible. The tendency for some motors to get stuck in slow rotation in
the wrong direction also seems to be fixed.

The only downside seems to be that the unbalanced filtering now seems to
have more tendancy to align to magnetic harmonics than actual motor
timing. This causes a squeak or squeal until the power level is changed
or the motor moves slightly, affecting the inductance. Traces on the
oscilloscope seem to look very much like those from a micro motor in this
case, so it seems difficult to avoid. Please report any such cases where
this is a problem.