Limit zero crossing filter backtracking, take 3.
Take 2 in d58be3e357 made some motors better and other cases, like the
MT3506, worse. It is a little sad how much logic is required, but I have
tried for weeks to find a more simple approach. Nothing else is able to
work as well as this with all of the motors I have to test.

Previously, the demagnetization path disabled power only during the zero
crossing wait. This did reduce the amount of current slightly, which did
help, but in many cases the timeout would keep recurring for some time,
reducing output power until the back-EMF was high enough. We now disable
power and align, which allows for faster acceleration even with timeouts.
The remaining timeouts can probably be solved by a smoother timing_duty
limit.

This also restores the 240 degree zero crossing timeout which helps with
motors with unbalanced windings and the ZMR RCX 1804 2400KV.