Disable (by default) rapid beeping at stop if invalid PWM pulses were received.
This is perhaps useful, but happens as a result of flight boards changing
timers or in other situations that probably make this more of a nuisance
than useful data. However, it is worth trying this option if sync loss is
suspected since it should indicate if the problem is related to PWM noise,
as in github issue #50.