Don't spray back at signals we can't follow.
The boot_rx4 loop was intending to fall through to boot_rx_bit, but got
sidelined by trampoline-avoiding interleaving with boot_tx, and so does
exactly that in the case of receiving crap or pulses too quickly from the
transmitter.

I was hitting this when trying to communicate at a 3 microsecond quarter
bit time (0-bit pulse length). The current reliable minimum speed seems
to be about 5 microseconds (20 microsecond bit time) which seems to be
mostly a result of jitter resulting from the timer prescaler not being
cleared when the timer counter is cleared, and late clearing. Boo.