Fast-path timing calculations and raise maximum speed to 312,500eRPM.
Previous maximum was about 178,571eRPM. Divide by "pole pairs" to find
actual RPM; for example, for a 12N14P motor, divide by (14 / 2) or 7.

The trade-off is slightly more code because of the two paths: one for
the usual full 24-bit calculations, and one for 16-bit calculations when
they are sufficient as a result of the shorter timing period.

Note that TIMING_MIN isn't even enforced since 48fe82e8 -- Oops.