Add note for new binary download location: http://0x.ca/tgy/downloads/
After over 25,605 downloads at https://github.com/sim-/tgy/downloads ,
it's clear that people are still confused where the latest version is.
I don't want to clutter the source with the binary outputs since there
are so many boards and variants. So, reference the other location in
the README.md.