enable sleep when armed in neutral
This implements sleep in armed but idle mode.
Some of the lables have been renamed to better describe what we do.
Some code consolidation has happened to minimize the code-size with this
new feature.

Now with USE_SLEEP set to 1 or 2 the measurements on a tgy.inc style board
give 40.2mA also when in armed but neutral mode.

This reduction may not seem worthwhile for all use-cases, but assume
a situation where we have 8 Motors on a ROV and only 2 of them are
running active, then you would waste 60mA (90mA when you leave
these ESC/motors unarmed)