sample ADC channels on a regular basis and oversample the result
This implements 256 x ADC oversampling with 2 bit decimation
to result in a result value range of [0:65472]

There is also some scaling offset calculation that can get applied
to the gathered values.

when USE_ADC = 1 is set then we look for defined:
* mux_temperature
* mux_voltage
* mux_current
and calculate the mask of adc channels to use.
Alternatively USE_ADC_MASK can be set to a bitmask of the ADC channels
to capture - note that there are no Channels 8 - 13, so if any of these
bits are set the build will fail .

For example with afro_nfet on can define:
.equ	USE_ADC = 1
.equ	mux_voltage_mul	= (5*256*(18000+3300)/3300)
to get the voltage scaled correctly to: Vbat*256

Note that typically an oversample of a channel takes:
0.1037s for a motor running at aprox 1400 commutation cycles/s
0.0282s for an unarmed ESC

note that when armed the sample rate may go as low as 1 second -
especially when braking or accelerating.
Also there is no sampling happening while the ESC is "beeping"!

Last thing missing is exporting those numbers via I2C...